refactor(symbols): Move private symbols into a util. Only expose public symbols via api.
