+++ src/api/symbols.js
@@ -1,8 +1,1 @@
-export const $created = '____skate_created';
-export const $ctor = '____skate_constructor';
-export const $events = '____skate_events';
-export const $name = '____skate_name';
-export const $props = '____skate_props';
-export const $renderer = '____skate_renderer';
-export const $rendererDebounced = '____skate_rendererDebounced';
-export const $shadowRoot = '____skate_shadowRoot';
+export { $name, $shadowRoot } from '../util/symbols';
